Added a few json-ignore to clean up output and changed structure

+++ examsec70/src/main/java/examsec70/examsec70/models/UsuarioLamina.java
@@ -27,5 +27,5 @@ public class UsuarioLamina {
     @JoinColumn(name = "lamId")
     private Lamina usuLamLamina;
 
-    private long cantidad;
+    private int cantidad;
 }
